Rewire backend fetch onto reqwest::blocking, simplify VCL-side request plumbing
VclBackend::get_response/BackendResp no longer bridge through the async
reqwest::Client + BgThread channel dance: Varnish invokes get_response/read
synchronously on a dedicated worker thread already, so they now call
reqwest::blocking::Client directly and BackendResp just forwards
reqwest::blocking::Response's std::io::Read impl. Request bodies go through
varnish-rs's new Ctx::req_body/req_body_state API instead of raw
VRB_Iterate/ObjIterate FFI calls.

The VCL-level client.init()/send() path is untouched (still async, still on
the shared BgThread runtime), but with RespMsg::Chunk gone (only ever used by
the now-removed streaming backend path), RespMsg collapses to a plain
Result<Response, Error> and process_req always fully buffers before sending.

Also:
- add client::get_req (mirrors get_resp) and use it from set_header/set_body/
  copy_headers_to_req instead of inline VclTransaction::Req matching
- split implementation.rs into reqwest_private/backend.rs (VCLBackend,
  BackendResp, probes) and reqwest_private/vcl_client.rs (process_req, Entry,
  Request, Response, VclTransaction, BgThread, the VCL-facing client methods)
- bump reqwest to 0.13 (rustls-tls -> rustls) and point varnish at a
  varnish-rs rev with the req_body API
