+++ Cargo.toml
@@ -9,7 +9,7 @@ publish = false
 [dependencies]
 anyhow = "1"
 bytes = "1"
-reqwest = { version = "0.12", features = ["stream", "deflate", "gzip", "brotli", "rustls-tls"] }
+reqwest = { version = "0.13", features = ["stream", "deflate", "gzip", "brotli", "rustls", "blocking"] }
 serde_json = "1"
 tokio = { version = "1", features = ["full"] }
 tokio-stream = "0.1.17"
